Temporary, before moving encoding to client
