+++ packages/common/src/components/TextboxClipboardWrapper/index.tsx
@@ -4,6 +4,7 @@ import Clipboard from 'clipboard';
 import { TextField, ITextField } from 'office-ui-fabric-react/lib/TextField';
 import { IconButton } from 'office-ui-fabric-react/lib/Button';
 import { invokeGlobalErrorHandler } from '../../utilities/splash.screen';
+import { generateCryptoSafeRandom } from '../../utilities/misc';
 
 interface IProps {
   text: string;
@@ -13,12 +14,14 @@ interface IProps {
 interface IState {}
 
 class TextboxClipboardWrapper extends React.Component<IProps, IState> {
+  private clipboardButtonClassName: string;
   private clipboard: Clipboard;
 
   constructor(props: IProps) {
     super(props);
 
-    this.clipboard = new Clipboard('.export-to-clipboard', {
+    this.clipboardButtonClassName = 'export-to-clipboard-' + generateCryptoSafeRandom();
+    this.clipboard = new Clipboard('.' + this.clipboardButtonClassName, {
       text: () => this.props.text,
     });
     this.clipboard.on('error', invokeGlobalErrorHandler);
@@ -40,7 +43,7 @@ class TextboxClipboardWrapper extends React.Component<IProps, IState> {
           <IconButton
             iconProps={{ iconName: 'Copy' }}
             ariaLabel="Copy to clipboard"
-            className="export-to-clipboard"
+            className={this.clipboardButtonClassName}
           />
         </OuterStyle>
       </>
