+++ packages/editor/src/utils/index.ts
@@ -114,14 +114,3 @@ const isCustomFunctionRegex = /@customfunction/i;
 export function isCustomFunctionScript(content: string) {
   return isCustomFunctionRegex.test(content);
 }
-
-export function isInternetExplorer(): boolean {
-  // From https://stackoverflow.com/a/19868056/678505
-  return 'ActiveXObject' in window;
-}
-
-export function generateCryptoSafeRandom(): number {
-  const random = new Uint32Array(1);
-  window.crypto.getRandomValues(random);
-  return random[0];
-}
